Изменена логика удаления пользователей администратором: запрещено удалять свой профиль админу

+++ server/app/api/v1/users/users_service.py
@@ -4,10 +4,10 @@ from uuid import UUID
 from fastapi import Depends
 
 from server.app.api.v1.auth.auth_manager import AuthManager
-from server.app.api.v1.common_schemas import CANT_CHANGE_OWN_ROLE_ERROR_TEXT
+from server.app.api.v1.common_schemas import CANT_CHANGE_OWN_ROLE_ERROR_TEXT, CANT_DELETE_OWN_PROFILE_ERROR_TEXT
 from server.app.api.v1.courses.courses import CourseResponse
 from server.app.api.v1.courses.courses_manager import CoursesManager
-from server.app.api.v1.notes.exceptions import CantChangeOwnRoleError
+from server.app.api.v1.notes.exceptions import CantChangeOwnRoleError, CantDeleteOwnProfileError
 from server.app.api.v1.users.users import (
     ApplicationById,
     ApplicationChangeStatus,
@@ -117,7 +117,9 @@ class UsersService:
             return await self.users_manager.change_role_to_professor(id=user.id)
         return await self.users_manager.change_role_except_professors(id=user.id, role=user.role)
 
-    async def delete_user(self, id: UUID) -> None:
+    async def delete_user(self, id: UUID, current_user_id: UUID) -> None:
+        if current_user_id == id:
+            raise CantDeleteOwnProfileError(CANT_DELETE_OWN_PROFILE_ERROR_TEXT)
         await self.users_manager.delete_user(id=id)
         await self.auth_manager.delete_sessions(user_id=id)
